% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateCCA.R
\name{AggregateCCA}
\alias{AggregateCCA}
\title{AggregateCCA -- a function that pulls the weights for each canonical model and variable and returns a data frame}
\usage{
AggregateCCA(CCAfold = CCAfold10, x_data = NULL, y_data = NULL, filename = NULL)
}
\arguments{
\item{CCAfold}{CCA variable output from CrossFoldCCA.R}

\item{x_data}{data frame containing the X data, if provided, will create weights for each CCA mode correlated against the X domain}

\item{y_data}{data frame containing the Y data, if provided, will create weights for each CCA mode correlated against the Y domain}

\item{filename}{character vector representing the filename for the csv output. If provided will create a csv with the output data frame.}

\item{threshold}{a positive number indicating the threshold to determine whether a variable is contributing}
}
\description{
This function is used as an input for wordcloudCCA.R to visualize outputs
}
\keyword{CCA,}
\keyword{wordcloud}
