% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyWB_to_data.R
\name{ApplyWB_to_data}
\alias{ApplyWB_to_data}
\title{ApplyWB_to_data -- generate bootstrapped data}
\usage{
ApplyWB_to_data(fit_val, resid_val, sample_val)
}
\arguments{
\item{fit_val}{Fitted values generated by ComputeFits(geeglm_obj)}

\item{resid_val}{Residual values generated by ComputeResiduals(geeglm_obj)}

\item{sample_val}{Sample values generated by ComputeMM_WB, or generated by the user}
}
\description{
This function will generate a wild bootstrap for a single vertex/voxel.
The function can be called in parallel to operate on a map. Such usage can be found in the main function
}
\examples{
WB_value <- ApplyWB_to_data(fit_val,resid_val,sample_val)
}
\keyword{bootstrap}
\keyword{wild}
