% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMarginalValue.R
\name{CalculateMarginalValue}
\alias{CalculateMarginalValue}
\title{CalculateMarginalValue -- a  function that will calculate a fitted value for a set of parameters}
\usage{
CalculateMarginalValue(beta_measures, marginal_measures)
}
\arguments{
\item{beta_measures}{The map of parameters for the fitted formula, including an intercept as the first parameter.}

\item{marginal_measures}{A numeric vector of measures to input per parameter. Ordered in the same order as parameters. Nothing should be specified for the first parameter}
}
\description{
This function is used in CalculateMarginalModel to calculate an intermediate map for visualizations
}
\examples{
marginal_value <- CalculateMarginalValue(beta_measures=beta_measures,marginal_measures=as.array(c(8,9,10)))
}
\keyword{estimator}
\keyword{fast}
\keyword{marginal}
\keyword{model}
\keyword{outputs}
\keyword{sandwich}
