% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePvalue.R
\name{CalculatePvalue}
\alias{CalculatePvalue}
\title{CalculatePvalue -- calculate the p value for cluster detection or via a lookup table}
\usage{
CalculatePvalue(observed_cc, WB_cc, nboot, sigtype)
}
\arguments{
\item{observed_cc}{the observed statistical value, can be the area of a connected component or the Z transformed Wald statistic}

\item{WB_cc}{The distribution of areal extents derived from the wild bootstrap -- used when performing cluster detection}

\item{nboot}{The number of wild bootstraps performed, used to calculate the p value for cluster detection}

\item{sigtype}{A character string representing the type of significance test, will perform cluster detection if "cluster" is specified.}
}
\description{
This function wraps many of the other functions to perform cluster detection analysis using
the wild bootstrap.
}
\examples{
p <- CalculatePvalue(observed_cc,WB_cc,nboot,sigtype)
}
\keyword{cluster}
\keyword{pvalues}
