% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBrainMetrics.R
\name{CiftiInputs}
\alias{CiftiInputs}
\title{LoadBrainMetrics -- A Modularlized function for MarginalModelCifti to load MR data}
\usage{
CiftiInputs(ciftilist, structtype)
}
\arguments{
\item{ciftilist}{A data frame comprising a list of path/filenames to load}

\item{structtype}{A character string denoting whether the map is volumetric ('volume'), surface-based ('surface'), a pconn ('pconn'), or a NIFTI connectivity matrix ('niiconn').}
}
\description{
This function loads the imaging data for ConstructMarginalModel. Created to further modularlize the R package.
}
\examples{
cifti_data <- LoadBrainMetrics(cifitilist,structtype)

}
\keyword{CIFTI}
\keyword{bootstrap}
\keyword{estimator}
\keyword{marginal}
\keyword{model}
\keyword{sandwich}
\keyword{scalar}
\keyword{wild}
