% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFastSwE.R
\name{ComputeFastSwE}
\alias{ComputeFastSwE}
\title{ComputeFastSwE -- a fast function to calculate errors via the sandwich estimator. Modified from Tom Nichols, Feb 1, 2019}
\usage{
ComputeFastSwE(X, nested, Nelm, resid_map, npredictors, beta_map,
  adjustment)
}
\arguments{
\item{X}{The predictor data recast as a matrix}

\item{nested}{A vector denoting the blocks (i.e. groups) which are used to calculate variance separately if set to NULL will compute as if a single group}

\item{Nelm}{An integer denoting the number of voxels/vertices}

\item{resid_map}{The residuals from the estimated data fits, usually by lm.fit or lmfit}

\item{npredictors}{The number of predictors in X}

\item{beta_map}{The beta map from the estimated data fits, usually by lm.fit or lmfit}

\item{adjustment}{The residuals will be adjusted according to the small sample size adjustment. Acceptable values are "HC2", "HC3", and NULL.}
}
\description{
This function is used to quickly calculate error from a marginal model via a sandwich estimator. The function can handle a single block (i.e. repeated measure)
}
\examples{
T_map <- ComputeFastSwE(X=external_df,nested=nested,Nelm=Nelm,resid_map=resid_map,npredictors=npredictors,beta_map=beta_map,adjustment=adjustment)
}
\keyword{estimator}
\keyword{fast}
\keyword{marginal}
\keyword{matrix}
\keyword{model}
\keyword{sandwich}
