% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFits.R
\name{ComputeFits}
\alias{ComputeFits}
\title{ComputeFits -- get estimated fits from a geeglm_object}
\usage{
ComputeFits(geeglm_obj, nmaps)
}
\arguments{
\item{geeglm_obj}{An GLM object that is produced from computing the marginal model}
}
\description{
This function will compute fits from a marginal model, which can be used to generate wild bootstraps.
The function can be called in parallel to operate on a map. Such usage can be found in the main function
}
\examples{
fits <- ComputeFits(geeglm_obj)
}
\keyword{bootstrapping}
\keyword{estimates}
