% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMM.R
\name{ComputeMM}
\alias{ComputeMM}
\title{ComputeMM -- compute the marginal model at a single vertex/voxel}
\usage{
ComputeMM(cifti_meas, external_df, notation, family_dist, corstr,
  zcor = NULL, wave = NULL, id_subjects)
}
\arguments{
\item{cifti_meas}{An array of N numeric measures, where N is the number of subjects}

\item{external_df}{A data frame with NxM measures. Should include all features that the user wants to model.}

\item{notation}{A character string representing the modelling formula. Uses the wilkinson notation.}

\item{family_dist}{A character string representing the type of distribution to use for modelling.}
}
\description{
This function will compute the marginal model on a single vertex or voxel. 
The function can be called in parallel to operate on a map. Such usage can be found in the main function
}
\examples{
geeglm_obj <- ComputeMM(cifti_meas,external_df,notation,family_dist)
}
\keyword{estimator}
\keyword{marginal}
\keyword{model}
\keyword{sandwich}
