% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMM_WB.R
\name{ComputeMM_WB}
\alias{ComputeMM_WB}
\title{ComputeMM_WB -- calculate the maximum cluster extent from a wild bootstrap}
\usage{
ComputeMM_WB(x, resid_map, fit_map, type, external_df, notation,
  family_dist, structtype, thresh, structfile, matlab_path, surf_command,
  corstr, wave, zcor, correctiontype, id_subjects, cifti_dim, nmeas, seeds)
}
\arguments{
\item{resid_map}{The residual map from the geeglm analysis}

\item{fit_map}{The fitted values map from the geeglm analysis}

\item{type}{A character string denoting the distribution to use for the wild bootstrap}

\item{external_df}{A data frame comprising non-brain measures to model.}

\item{notation}{A character string representing the model formula. Uses Wilkinson notation.}

\item{family_dist}{A character string denoting the assumed underlying distribution of all input data}

\item{structtype}{A character string denoting whether the map is volumetric ('volume') or surface-based}

\item{matlab_path}{A character string denoting the path to the matlab compiler. Please use v91.}

\item{surf_command}{A character string denoting the path to the SurfConnectivity command.}

\item{correctiontype}{A character string denoting cluster ('cluster') or point ('point') comparisons}

\item{cifti_map}{surface array or nifti containing the scalar values -- just used to determine the number of participants}

\item{zscore_map}{The z-scored map from the geeglm analysis}
}
\description{
This function wraps many of the other functions to perform cluster detection analysis using
the wild bootstrap.
}
\examples{
max_cc <- ComputeMM_WB(cifti_map,zscore_map,resid_map,fit_map,type,external_df,
notation,family_dist,structtype,thresh,structfile,matlab_path,surf_command,correctiontype)
}
\keyword{bootstrap}
\keyword{wild}
