% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeResiduals.R
\name{ComputeResiduals}
\alias{ComputeResiduals}
\title{ComputeResiduals -- get residuals from a geeglm_object}
\usage{
ComputeResiduals(geeglm_obj)
}
\arguments{
\item{geeglm_obj}{An GLM object that is produced from computing the marginal model}
}
\description{
This function will compute residuals from a marginal model, which can be used to generate wild bootstraps.
The function can be called in parallel to operate on a map. Such usage can be found in the main function
}
\examples{
resids <- ComputeResiduals(geeglm_obj)
}
\keyword{bootstrapping}
\keyword{residuals}
