% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeZscores.R
\name{ComputeZscores}
\alias{ComputeZscores}
\title{ComputeZscores -- convert estimated Wald statistics to z scores}
\usage{
ComputeZscores(geeglm_obj, nmaps)
}
\arguments{
\item{geeglm_obj}{An GLM object that is produced from computing the marginal model}
}
\description{
This function will compute z scores from Wald statistics, which can be used to infer statistical significance.
The function can be called in parallel to operate on a map. Such usage can be found in the main function
}
\examples{
zscores <- ComputeZscores(geeglm_obj)
}
\keyword{normalization}
