% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructMarginalModel.R
\name{ConstructMarginalModel}
\alias{ConstructMarginalModel}
\title{ConstructMarginalModel -- the main function that will estimate the significance of the marginal model fitted to a CIFTI dscalar dataset}
\usage{
ConstructMarginalModel(external_df, concfile, structtype, structfile,
  matlab_path, surf_command, wave = NULL, notation, zcor = NULL,
  corstr = NULL, family_dist = NULL, dist_type = "radenbacher",
  z_thresh = 2.3, nboot = 1000, p_thresh = 0.5,
  sigtype = "cluster", id_subjects = "subid",
  output_directory = "~/", ncores = 1, fastSwE = TRUE,
  adjustment = NULL, norm_external_data = TRUE,
  norm_internal_data = TRUE, marginal_outputs = FALSE,
  marginal_matrix = NULL, enrichment_path = NULL, modules = NULL,
  wb_command = "wb_command", subsetfile = NULL,
  permutation_directory = NULL)
}
\arguments{
\item{external_df}{A data frame comprising non-brain measures to model. Can be specified as a string to a csv file with appropriate headers.}

\item{concfile}{A character string denoting a single column text file that lists the dscalars in the same order as the external_df and wave frames.}

\item{structtype}{A character string denoting whether the map is volumetric ('volume'), surface-based ('surface'), a pconn ('pconn'), or a NIFTI connectivity matrix ('niiconn').}

\item{structfile}{A character string denoting the structural map file, used for cluster detection on surfaces only.}

\item{matlab_path}{A character string denoting the path to the matlab compiler. Please use v91.}

\item{surf_command}{A character string denoting the path to the SurfConnectivity command.}

\item{wave}{A data frame comprising the order of the waves (longitudinal measures) for each independent case.}

\item{notation}{A character string representing the model formula. Uses Wilkinson notation.}

\item{zcor}{A matrix defining the correlation structure set by the user, can be used to specify custom dependencies (e.g. site or kinship matrices).}

\item{corstr}{A character string that defines the correlation structure in a predetermined way. "Independence" is recommend for most cases.}

\item{family_dist}{A character string denoting the assumed underlying distribution of all input data.}

\item{dist_type}{A character string denoting the distribution to use for the wild bootstrap.}

\item{z_thresh}{A numeric that represents the threshold for Z statistics when performing cluster detection.}

\item{nboot}{A numeric that represents the number of wild bootstraps to perform.}

\item{p_thresh}{A numeric that represents the p value threshold for assessing significance.}

\item{sigtype}{A character string denoting cluster ('cluster'),  point ('point'), or enrichment ('enrichment') comparisons.}

\item{id_subjects}{A character string denoting the column for the subject ID. Only needed when FastSwE is FALSE}

\item{output_directory}{A character string denoting the path to the output MRI statistical maps}

\item{ncores}{An integer denoting the number of CPU cores to use when conducting permutation tests}

\item{adjustment}{A character string denoting the small sample size adjustment to use when fastSwE is set to TRUE. Is NULL by default.}

\item{norm_external_data}{A boolean. If set to true, external data will be normed prior to analysis.}

\item{norm_internal_data}{A boolean. If set to true, MRI data will be normed per datapoint prior to analysis.}

\item{marginal_outputs}{A boolean. If set to true, marginal values will be output as statistical maps}

\item{marginal_matrix}{A numeric matrix depicting how to draw the map. Only needed if marginal_outputs is set to TRUE}

\item{enrichment_path}{A string depicting the path to the enrichment code. Used for enrichment analysis only (i.e. when sigtype is set to 'enrichment').}

\item{modules}{A csv file or array that depicts the modules for the enrichment analysis.}

\item{wb_command}{A character string denoting the path to the wb_command file}

\item{subsetfile}{A character string denoting the path to a subset selection file, to select only a subset of the participants}

\item{permutation_directory}{A character string denoting the path to the directory containing ONLY permutation tests. If set to NULL, permutation tests will be generated on the fly}

\item{fastSWE}{A boolean that determine the sandwhich estimator approach. If set to FALSE, will use standard R package geeglm. If set to TRUE will use custom-built estimator using rfast.}
}
\description{
This function wraps all of the other functions in the MarginalModelCifti package. The output is a significance map, which is either uncorrected or cluster detection was performed.
}
\examples{
max_cc <- ComputeMM_WB(cifti_map,zscore_map,resid_map,fit_map,type,external_df,
notation,family_dist,structtype,thresh,structfile,matlab_path,surf_command,correctiontype)
}
