% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertVolume.R
\name{ConvertVolume}
\alias{ConvertVolume}
\title{ConvertVolume -- convert a 3D MRI volume into a list of values}
\usage{
ConvertVolume(vol_array, vol_dim)
}
\arguments{
\item{vol_array}{A 3D array that contains values per voxel}

\item{vol_dim}{A numeric of three values to use to identify the size of each dimension (X,Y,Z)}
}
\description{
This function will return a 3D array for items within a list. The dimensions of the image must be
specified and match the same number of items as the list. Useful when converting an array to a list
See below for examples.
}
\examples{
thresh_vol <- ConvertVolume(thresh_map,vol_dim)
}
\keyword{conversion}
\keyword{volume}
