% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineNestedGroups.R
\name{DetermineNestedGroups}
\alias{DetermineNestedGroups}
\title{DetermineNestedGroups -- takes mulitiple within/between subject measures to create unique blocks for calculating SwE}
\usage{
DetermineNestedGroups(wave)
}
\arguments{
\item{wave}{A data frame typically read from a csv file.}
}
\description{
This function will identify unique groups based on the input wave dataset. 
These groups will have their covariance structure calculated separately for the SwE
See below for examples.
}
\examples{
final_wave <- DetermineNestedGroups(wave)
}
\keyword{blocks}
\keyword{wave}
