% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentAnalysis.R
\name{EnrichmentAnalysis}
\alias{EnrichmentAnalysis}
\title{WriteMatrixToCifti -- Output an R array matrix into an appropriate CIFTI format}
\usage{
EnrichmentAnalysis(metric_data, ncols, modules, matlab_path,
  enrichment_path, output_file)
}
\arguments{
\item{metric_data}{A binarized matrix}

\item{ncols}{The number of columns in the matrix -- used to write the array to a text file}

\item{modules}{A numeric vector representing the modules for the binarized modules}

\item{matlab_path}{A character string representing the path to the matlab compiler}

\item{enrichment_path}{A character string representing the path to the compiled enrichment folder}

\item{output_file}{A character string representing the path to the output chisquared file}
}
\description{
This function will write an R data matrix as a CIFTI pconn file, properly indexed.
Unfortunately, this function requires a compiled matlab package, SurfConnectivity.
Contact Eric Feczko to get the package.
}
\examples{
WriteMatrixToCifti(metric_data,ncols,surf_template_file,matlab_path,enrichment_path,output_file)
}
\keyword{cluster}
\keyword{surface}
