% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSurfAreas.R
\name{GetSurfAreas}
\alias{GetSurfAreas}
\title{GetSurfAreas -- Get computed size of connected surface components}
\usage{
GetSurfAreas(metric_data, surf_file, matlab_path, surf_command)
}
\arguments{
\item{metric_data}{A numeric array containing only binary values (e.g. 0;1)}

\item{surf_file}{A character string representing the path to the surface file}

\item{matlab_path}{A character string representing the path to the matlab compiler}

\item{surf_command}{A character string representing the path to the compiled surface command}
}
\description{
This function will compute the size of connected components in statistical surfaces.
Unfortunately, this function requires a compiled matlab package, SurfConnectivity.
Contact Eric Feczko to get the package.
}
\examples{
component_sizes <- GetSurfAreas(metric_data,surf_file,matlab_path,surf_command)
}
\keyword{cluster}
\keyword{surface}
