% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVolAreas.R
\name{GetVolAreas}
\alias{GetVolAreas}
\title{GetVolAreas -- Get computed size of connected volumetric components}
\usage{
GetVolAreas(bin_image)
}
\arguments{
\item{bin_image}{A nifti array containing only binary values (e.g. 0;1)}
}
\description{
This function will compute the size of connected components in statistical volumes
}
\examples{
component_sizes <- GetVolAreas(bin_image)
}
\keyword{cluster}
\keyword{volume}
