% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix2Vector.R
\name{Matrix2Vector}
\alias{Matrix2Vector}
\title{Matrix2Vector -- a function that will map a matrix to a vector and vice versa}
\usage{
Matrix2Vector(pconn_data = NULL, pconn_vector = NULL,
  direction = "to_vector")
}
\arguments{
\item{pconn_data}{The pconn matrix represented as a 2D matrix}

\item{pconn_vector}{The pconn vector to map to the matrix. Only used when "to_matrix" is specified}

\item{direction}{A string denoting the direction of the transform. Can be either "to_matrix" or "to_vector". Default is "to_vector"}
}
\description{
This function is used when loading pconns in ConstructMarginalModel.
}
\examples{
data <- Matrix2Vector(pconn_data=cifti_data$data,pconn_vector=NULL,direction="to_vector")
data <- Matrix2Vector(pconn_data=cifti_data$data,pconn_vector=pvalue_vector,direction="to_matrix")
}
\keyword{cifti}
\keyword{marginal}
\keyword{model}
\keyword{pconn}
