% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseDf.R
\name{ParseDf}
\alias{ParseDf}
\title{ParseDf -- a fast function to extract data from a data frame and convert to a design matrix. Used when fastSwE is set to true}
\usage{
ParseDf(external_df, notation)
}
\arguments{
\item{external_df}{The data frame used to extract data}

\item{notation}{The formula specified by the user}
}
\description{
This function is used to extract data from a data frame and convert to a design matrix. The matrix is then input into SwE for fast computations.
}
\examples{
X <- ParseDf(external_df,notation)
}
\keyword{estimator}
\keyword{fast}
\keyword{marginal}
\keyword{matrix}
\keyword{model}
\keyword{sandwich}
