% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepCIFTI.R
\name{PrepCIFTI}
\alias{PrepCIFTI}
\title{PrepCIFTI -- A function for laoding cifti data directly. Not useful for cluster detection}
\usage{
PrepCIFTI(cifti_file)
}
\arguments{
\item{cifti_file}{A character string representing the path to a cifti file.}
}
\description{
This function is a simple wrapper for reading cifti data
}
\examples{
cifti_data <- PrepCIFTI(cifti_file)
cifti_dataset <- lapply(cifti_file,PrepCIFTI)
}
\keyword{cifti}
\keyword{scalar}
