% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepSurf.R
\name{PrepSurf}
\alias{PrepSurf}
\title{PrepSurf -- A function for loading gifti surfaces}
\usage{
PrepSurf(surf_file)
}
\arguments{
\item{surf_file}{A character string representing the path to a gifti file.}
}
\description{
This function is a simple wrapper for read_gifti.
}
\examples{
surface_data <- PrepSurf(gifti_file)
surfaces_data <- lapply(gifti_filelist,PrepSurf)
}
\keyword{gifti}
\keyword{surface}
