% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepSurfMetric.R
\name{PrepSurfMetric}
\alias{PrepSurfMetric}
\title{PrepSurfMetric -- A function for loading gifti metric data}
\usage{
PrepSurfMetric(funcgii_file)
}
\arguments{
\item{funcgii_file}{A character string representing the path to a func.gii file.}
}
\description{
This function is a simple wrapper for reading gifti metric data
}
\examples{
gifti_data <- PrepSurfMetric(giifunc_file)
gifti_dataset <- lapply(giifunc_filelist,PrepSurfMetric)
}
\keyword{gifti}
\keyword{metric}
