% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepVolMetric.R
\name{PrepVolMetric}
\alias{PrepVolMetric}
\title{PrepVolMetric -- A function for loading nifti volume metric data}
\usage{
PrepVolMetric(vol_file)
}
\arguments{
\item{vol_file}{A character string representing the path to a gifti file.}
}
\description{
This function is a simple wrapper for reading nifti metric volumes
}
\examples{
niftifile <- PrepVolMetric(nifti_file)
niftifiles <- lapply(nifti_filelist,PrepVolMetric)
}
\keyword{nifti}
\keyword{volume}
