% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReframeCIFTIdata.R
\name{ReframeCIFTIdata}
\alias{ReframeCIFTIdata}
\title{ReframeCIFTIdata -- converts CIFTI scalar data from multiple subjects into a greyordinate x subject data frame that is easier to manipulate}
\usage{
ReframeCIFTIdata(index, cifti_rawmeas)
}
\arguments{
\item{index}{the greyordinate to extract and convert into a data frame}

\item{cifti_rawmeas}{The object created by PrepCIFTI, when multiple subjects are loaded (e.g.  "transpose(data.frame((lapply(as.character(ciftilist$file),PrepCIFTI))))")}
}
\description{
This function wraps many of the other functions to perform cluster detection analysis using
the wild bootstrap.
}
\examples{
cifti_meas <- ReframeCIFTIData(index,cifti_rawmeas)
}
\keyword{CIFTI}
\keyword{import}
\keyword{scalar}
