% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevertVolume.R
\name{RevertVolume}
\alias{RevertVolume}
\title{RevertVolume -- convert a list of values into a 3D MRI volume.}
\usage{
RevertVolume(vol_list, vol_dim)
}
\arguments{
\item{vol_list}{A list item that contains values per voxel}

\item{vol_dim}{A numeric of three values to use to identify the size of each dimension (X,Y,Z)}
}
\description{
This function will return a 3D array for items within a list. The dimensions of the image must be
specified and match the same number of items as the list. Useful when converting an array to a list
See below for examples.
}
\examples{
thresh_vol <- RevertVolume(thresh_map,vol_dim)
}
\keyword{conversion}
\keyword{volume}
