% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveBWASfile.R
\name{SaveBwasfile}
\alias{SaveBwasfile}
\title{SaveBWASfile -- A Modularlized function for MarginalModelCifti to save statistical outputs}
\usage{
SaveBwasfile(
  BWAS_statmap,
  structtype,
  output_prefix,
  measnames,
  surf_command = NULL,
  wb_command = NULL,
  zeros_array = NULL,
  matlab_path = NULL,
  sigtype = NULL
)
}
\arguments{
\item{BWAS_statmap}{A matrix containing a brain-wide assocation statistical map, can be anything as long as it matches the dimensions of the inputs}

\item{structtype}{A character string denoting whether the map is volumetric ('volume'), surface-based ('surface'), a pconn ('pconn'), or a NIFTI connectivity matrix ('niiconn').}

\item{output_prefix}{A character string depicting the prefix of the output filenames, usually depicting the map metric}

\item{measnames}{A list of measures (covariates) being tested by the model, used to label the maps appropriately.}

\item{surf_command}{A path to the SurfConnectivity toolbox used to perform CIFTI/GIFTI writing}

\item{wb_command}{A full path/filename to thte wb_command file, used for ciftis}

\item{zeros_array}{A 2D array filled with zeros representing the initial structure for pconns and niicons, used to speed up file-saving}

\item{matlab_path}{A full path/filename to the matlab runtime compiler engine, must be V91}

\item{sigtype}{A character string denoting whether the map is volumetric ('volume'), surface-based ('surface'), a pconn ('pconn'), or a NIFTI connectivity matrix ('niiconn').}
}
\description{
This function saves stat map data for ConstructMarginalModel. Created to further modularlize the R package.
}
\examples{
SaveBWASfile(BWAS_statmap,structtype,output_prefix,measnames)

}
\keyword{CIFTI}
\keyword{bootstrap}
\keyword{estimator}
\keyword{marginal}
\keyword{model}
\keyword{sandwich}
\keyword{scalar}
\keyword{wild}
