% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSelection.R
\name{SubSelection}
\alias{SubSelection}
\title{SubSelection -- A function that returns indices for subselecting datasets}
\usage{
SubSelection(sublist, subsetlist)
}
\arguments{
\item{sublist}{A data frame comprising non-brain measures to model. Can be specified as a string to a csv file with appropriate headers.}

\item{subsetlist}{A character string denoting a single column text file that lists the dscalars in the same order as the external_df and wave frames.}
}
\description{
This function wraps all of the other functions in the MarginalModelCifti package. The output is a significance map, which is either uncorrected or cluster detection was performed.
}
\examples{
list_bin <- SubSelection(sublist,subsetlist)
}
