% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreshMap.R
\name{ThreshMap}
\alias{ThreshMap}
\title{ThreshMap -- threshold a list of values and return a binary list, can handle multiple values within each item of a list}
\usage{
ThreshMap(zscore_val, zthresh)
}
\arguments{
\item{zscore_val}{A list item that contains values}

\item{z_thresh}{A value to use to threshold the image}
}
\description{
This function will return a logical value for items within a list, used with purrr's map function, can be used to
parallelize thresholding on a list of many values (e.g. from CIFTIs). See below for examples.
}
\examples{
thresh <- ThreshMap(zscore_val,z_thresh)
}
\keyword{thresholding}
