% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteMatrixToCifti.R
\name{WriteMatrixToCifti}
\alias{WriteMatrixToCifti}
\title{WriteMatrixToCifti -- Output an R array matrix into an appropriate CIFTI format}
\usage{
WriteMatrixToCifti(
  metric_data,
  ncols,
  surf_template_file,
  matlab_path,
  surf_command,
  output_file,
  wb_command
)
}
\arguments{
\item{ncols}{The number of columns in the matrix -- used to write the array to a text file}

\item{surf_template_file}{A character string representing the path to the surface template file, only used for indexing.}

\item{matlab_path}{A character string representing the path to the matlab compiler}

\item{surf_command}{A character string representing the path to the compiled SurfConnectivity folder}

\item{output_file}{A character string representing the path to the output gifti file}

\item{metric_data_matrix}{A numeric vector containing any values. Indexing must match the template file}
}
\description{
This function will write an R data matrix as a CIFTI pconn file, properly indexed.
Unfortunately, this function requires a compiled matlab package, SurfConnectivity.
Contact Eric Feczko to get the package.
}
\examples{
WriteMatrixToCifti(metric_data,ncols,surf_template_file,matlab_path,surf_command,output_file)
}
\keyword{cluster}
\keyword{surface}
