% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteVectorToGifti.R
\name{WriteVectorToGifti}
\alias{WriteVectorToGifti}
\title{WriteVectorToGifti -- Output an R data vector into an appropriate GIFTI format}
\usage{
WriteVectorToGifti(metric_data, surf_template_file, matlab_path,
  surf_command, output_file)
}
\arguments{
\item{metric_data}{A numeric vector containing any values. Indexing must match the template file}

\item{surf_template_file}{A character string representing the path to the surface template file, only used for indexing.}

\item{matlab_path}{A character string representing the path to the matlab compiler}

\item{surf_command}{A character string representing the path to the compiled SurfConnectivity folder}

\item{output_file}{A character string representing the path to the output gifti file}
}
\description{
This function will write an R data vector as a GIFTI file, properly indexed.
Unfortunately, this function requires a compiled matlab package, SurfConnectivity.
Contact Eric Feczko to get the package.
}
\examples{
WriteVectorToGifti(metric_data,surf_template_file,matlab_path,surf_command,output_file)
}
\keyword{cluster}
\keyword{surface}
