% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepNiiConnMetric.R
\name{PrepNiiConnMetric}
\alias{PrepNiiConnMetric}
\title{PrepNiiConnMetric -- A function for loading nifti connectivity metric data}
\usage{
PrepNiiConnMetric(niiconn_file)
}
\arguments{
\item{niiconn_file}{A character string representing the path to a nifti connectivity matrix file.}
}
\description{
This function is a simple wrapper for reading connectivity matrices from niftis. 
The function will check to see whether the qform is needed before loading in the data.
Generally, the qform shouldn't be needed, if it is used, there may be a warning when using the dataset.
}
\examples{
niftifile <- PrepNiiConnMetric(nifti_file)
niftifiles <- lapply(nifti_filelist,PrepNiiConnMetric)
}
\keyword{conn}
\keyword{nifti}
\keyword{volume}
